requirements.txt
streamlit==1.29.0
pillow==10.1.0
requests==2.31.0
alibabacloud_alimt20181012==1.5.2
alibabacloud_tea_openapi==0.3.16
alibabacloud_tea_util==0.3.13
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
apscheduler==3.10.4
sse-starlette==1.6.1
sentry-sdk==1.40.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
